% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prediction_correction.R
\name{add_missing_first_month}
\alias{add_missing_first_month}
\title{Fonction utilitaire: corriger l'absence du mois de janvier 2015}
\usage{
add_missing_first_month(df)
}
\arguments{
\item{df}{\code{data.frame()}}
}
\value{
\code{data.frame()}
}
\description{
Le mois de janvier est imputé par la moyenne des autresmois de la période.
}
