% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prediction_correction.R
\name{adjust_seasonnality}
\alias{adjust_seasonnality}
\title{Correction de la saisonnalité pour le secteur 000C1}
\usage{
adjust_seasonnality(df, n_month)
}
\arguments{
\item{df}{\code{data.frame()} Données agrégées par périodes.}

\item{n_month}{\code{int()} Durée de la plage d'agrégation.}
}
\value{

}
\description{
Une forte saisonnalité a été observé pour le secteur 000C1.
Cette fonction corrige cette saisonnalité.
}
