% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sector_correction.R
\name{aggregate_by_n_months}
\alias{aggregate_by_n_months}
\title{Agrège les donnée de défaillances par plages de plusieurs mois}
\usage{
aggregate_by_n_months(df, n_month = 3)
}
\arguments{
\item{df}{\code{data.frame} Données à agréger, avec colonnes "count",
"count_new_outcome", "secteur", "nom_secteur"}

\item{n_month}{\verb{1, 2, 3, 6} durée de la plage en mois}
}
\value{
\code{data.frame} Données agrégées, "count" est moyenné,
"count_new_outcome" est sommé, "prop_new_outcome" est recalculé.
}
\description{
Agrège les donnée de défaillances par plages de plusieurs mois
}
