% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prediction_correction.R
\name{aggregate_by_secteur}
\alias{aggregate_by_secteur}
\title{Agrège les données de défaillances par secteur}
\usage{
aggregate_by_secteur(df)
}
\arguments{
\item{df}{\code{data.frame()} Données agrégées.}
}
\value{
\code{data.frame()}
}
\description{
Agrège les données de défaillances par secteur
}
