% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{aggregate_explainer_frame}
\alias{aggregate_explainer_frame}
\title{Agrège les facteurs d'importance selon une colonne d'une matrice
d'agrégation}
\usage{
aggregate_explainer_frame(frame, aggregation_matrix, group_name)
}
\arguments{
\item{frame}{\code{data.frame()} \cr
Table d'importance comme produite en sortie de xgb.importance.}

\item{aggregation_matrix}{\code{dataframe()} \cr
Table de correspondance qui possède une colonne "variable" et une colonne
correspondant au paramètre \code{group_name}. Si égal à NULL (défaut), alors les variables ne
sont pas aggrégées.}

\item{group_name}{\code{character()} \cr Nom de la colonne de
\code{aggregation_matrix} qui sert pour l'aggregation.}
}
\description{
Agrège les facteurs d'importance selon une colonne d'une matrice
d'agrégation
}
