% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{aggregate_importance_frame}
\alias{aggregate_importance_frame}
\title{Agrège les facteurs d'importance selon une colonne d'une matrice
d'agrégation}
\usage{
aggregate_importance_frame(frame, aggregation_matrix, group_name)
}
\arguments{
\item{frame}{\code{data.frame()} \cr
Table d'importance comme produite en sortie de xgb.importance.}

\item{aggregation_matrix}{\code{dataframe()} \cr
Table de correspondance qui possède une colonne "variable" et une colonne
correspondant au paramètre \code{group_name}. Si égal à NULL (défaut), alors les variables ne
sont pas aggrégées. Utiliser \code{read_aggregation_matrix()} pour lire la
table de corrsepondance.}

\item{group_name}{\code{character()} \cr Nom de la colonne de
\code{aggregation_matrix} qui sert pour l'agrégation.}
}
\description{
Les variables absentes de la matrice d'agrégation sont automatiquement
}
