% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{aggregate_local_explainer}
\alias{aggregate_local_explainer}
\title{Agrège les facteurs d'importance selon une colonne d'une matrice
d'agrégation}
\usage{
aggregate_local_explainer(dt_frame, aggregation_matrix, group_name)
}
\arguments{
\item{aggregation_matrix}{\code{dataframe()} \cr
Table de correspondance qui possède une colonne "variable" et une colonne
correspondant au paramètre \code{group_name}. Si égal à NULL (défaut), alors les variables ne
sont pas aggrégées. Utiliser \code{read_aggregation_matrix()} pour lire la
table de corrsepondance.}

\item{group_name}{\code{character()} \cr Nom de la colonne de
\code{aggregation_matrix} qui sert pour l'agrégation.}
}
\description{
Agrège les facteurs d'importance selon une colonne d'une matrice
d'agrégation
}
