% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{apply_corrections}
\alias{apply_corrections}
\title{Applique des corrections à la prédiction de l'apprentissage automatique}
\usage{
apply_corrections(task, correction_debt, correction_sector)
}
\arguments{
\item{correction_debt}{\code{data.frame} Corrections liées aux entreprises ayant
des débits sur les cotisations sociales. Colonnes "siret" et "correction_debt".}

\item{correction_sector}{\code{data.frame} Corrections liées aux secteurs
d'activité selon qu'ils soient plus ou moins touchés par la crise.
Colonnes "secteur" et "correction_sector".}
}
\value{
\code{sf_task} avec nouveau champ "full_prediction", un data frame avec
les colonnes: "siret", "periode", la décomposition en variables latentes,
la prédiction de l'apprentissage automatique, les corrections apportées,
la prédicition corrigé (tout ça dans l'espace des log-vraisemblance) et la
correction corrigée dans l'espace des probabilités
}
\description{
Après avoir prédit sur de nouvelles données, certaines corrections a
posteriori sont appliquées à la prédiction. Les corrections se présentent
comme des termes supplémentaires dans l'espace des log-vraisemblances.
}
