% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{apply_fte_map}
\alias{apply_fte_map}
\title{Préparation des données pour l'entraînement}
\usage{
apply_fte_map(data_to_prepare, is_train_frame, options)
}
\arguments{
\item{data_to_prepare}{\code{data.frame()} \cr Données à préparer}

\item{shape_function}{\verb{function(data.frame => ?)} \cr
Fonction qui met en forme les données pour l'apprentissage.}
}
\value{
\code{list(2)} \cr
Une liste avec deux champs:
\itemize{
\item "data": \code{H2OFrame}; les données préparées
\item "preparation_map": \code{list(H2OFrame)}; la carte de target encoding
}
}
\description{
Prépare les données pour l'entraînement, avec du target encoding et la
transformation en matrice pour alimenter xgboost.
}
\details{
La carte de target encoding est retourné en paramètre de sortie afin de
pouvoir l'utiliser sur d'autres échantillons de données.
}
