% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{assert_only_valid_columns}
\alias{assert_only_valid_columns}
\title{Check if any column is character or factor}
\usage{
assert_only_valid_columns(data_to_shape)
}
\description{
Check if any column is character or factor
}
