% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{assert_preparation_map}
\alias{assert_preparation_map}
\title{Asserts that preparation map is present in data}
\usage{
assert_preparation_map(task)
}
\description{
Asserts that preparation map is present in data
}
