% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_urssaf_evaluation.R
\name{aux_custom_eval_urssaf}
\alias{aux_custom_eval_urssaf}
\title{Fonction d'évaluation de survenue de nouvelles dettes urssaf.}
\usage{
aux_custom_eval_urssaf(eval_frame, F1_thres, F2_thres)
}
\arguments{
\item{eval_frame}{\code{data.frame()} \cr Un data.frame à évaluer comme produit par
l'objet \code{MLsegmentr::Assesser}. Se référer à la documentation
TODO: référence doc MLsegmentr 5fb65cac-b478-4413-a359-294e647103e3}

\item{F1_thres}{Seuil qui maximise le score F1, calculés sur
les données d'entraînement.}

\item{F2_thres}{Seuil qui maximise le score F2.}
}
\value{
\code{data.frame} \cr avec en plus des colonnes classiques "model",
"target_type", "segment"; les colonnes "statut" (niveau d'alerte), "count"
(nombre d'entreprises); "moyenne" (taux de défaillance moyen), "lconf" et
"hconf" (limites basse et haute respectivement de l'intervalle de confiance
issue d'une loi binomiale, à 95%)
}
\description{
Cette fonction est prévue pour construire un objet de type
\code{MLsegmentr::eval_function}, afin d'être utilisé dans un objet
\code{MLsegmentr::Assesser}. À utiliser avec les options par défaut
wrap_eval_fun = TRUE et unnest = TRUE.
}
