% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_urssaf_evaluation.R
\name{aux_custom_plot_urssaf}
\alias{aux_custom_plot_urssaf}
\title{Fonction graphique pour la précision}
\usage{
aux_custom_plot_urssaf(eval_frame)
}
\arguments{
\item{eval_frame}{\code{data.frame()} \cr Table évaluée tel que donné par la fonction
\code{[aux_custom_eval_urssaf]} avec paramètres wrap_eval_fun et unnest = TRUE.}
}
\description{
Cette fonction est prévue pour construire un objet EvaluationFunctions
afin d'être utilisé dans un objet Assesser du package MLsegmentr. À
utiliser avec les options par défaut wrap_eval_fun = TRUE et unnest = TRUE.
}
\details{
Elle s'utilise en complément de aux_custom_eval_urssaf, donc elle prend le
résultat en variable d'entrée.
}
