% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sector_correction.R
\name{check_if_same_string}
\alias{check_if_same_string}
\title{Fonction utilitaire}
\usage{
check_if_same_string(strings)
}
\arguments{
\item{strings}{\code{character} vecteur de chaînes de caractères}
}
\description{
Si toutes les chaînes sont identiques, renvoie la valeur
identique, sinon renvoie "*".
}
