% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{check_overwrites}
\alias{check_overwrites}
\title{Vérification de champs}
\usage{
check_overwrites(task, field_names)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{field_names}{\code{character()} \cr Nom des champs à vérifier.}
}
\value{
Nom des champs écrasés, \code{character(0)} sinon.
}
\description{
Vérifie si les champs qui vont être écrits sont déjà existant, et le cas
échéant vont être écrasés.
}
