% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_debt_correction.R
\name{compute_debt_correction}
\alias{compute_debt_correction}
\title{Calcule une correction pour les entreprises endettées}
\usage{
compute_debt_correction(task)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}
}
\value{
\code{data.frame} avec colonnes "siret" et "correction_debt"
}
\description{
La correction est dans l'espace des log-vraisemblance (donc après avoir
appliqué un logit aux prédictions en probabilité).
}
