% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sector_correction.R
\name{compute_sectorial_correction}
\alias{compute_sectorial_correction}
\title{Calcule la correction sectorielle liée à la crisé}
\usage{
compute_sectorial_correction(
  task,
  conjuncture_predictions = get_conjuncture_predictions(task)
)
}
\arguments{
\item{conjuncture_predictions}{\code{data.frame()} Predictions sectorielles, par
plages temporelles, calculés à partir des enquêtes de conjoncture. Format de retour de la
fonction \code{get_conjuncture_predictions}.}
}
\value{
\code{data.frame} avec colonnes "secteur" et "correction_sector"
}
\description{
La correction est dans l'espace des log-vraisemblance (donc après avoir
appliqué un logit aux prédictions en probabilité).
}
