% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_database.R
\name{connect_to_h2o}
\alias{connect_to_h2o}
\alias{connect_to_spark}
\title{Wrapper pour se connecter à H2O avec la bonne configuration}
\usage{
connect_to_h2o()

connect_to_spark(database = NULL, collection = NULL)
}
\description{
Créé une instance ou rejoint l'instance existante le cas échéant.
}
