% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{consolidate}
\alias{consolidate}
\title{Consolidate several tasks into a frame for evaluation}
\usage{
consolidate(..., data_name)
}
\arguments{
\item{...}{\code{sf_tasks | cv_tasks} \cr Tasks to consolidate}

\item{data_name}{\code{character()} \cr Name of the type of data to consolidate
("train_data", "validation_data", "test_data")}
}
\description{
Consolidate several tasks into a frame for evaluation
}
