% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_h2o}
\alias{convert_to_h2o}
\title{Conversion d'un data.frame en H2OFrame}
\usage{
convert_to_h2o(table_to_convert)
}
\arguments{
\item{table_to_convert}{\code{data.frame()} \cr Table à convertir.}
}
\value{
\code{h2o::H2OFrame} \cr
La table convertie.
}
\description{
Convertit un data.frame en H2OFrame en s'assurant de conserver les bons
types de données.
}
