% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{create_fte_pipeline}
\alias{create_fte_pipeline}
\title{Creates a PipeOp for impact encoding}
\usage{
create_fte_pipeline(target_encode_fields)
}
\arguments{
\item{target_encode_fields}{\code{character()} Name of fields to "feature target
encode (fte)" (or "impact encode")}
}
\value{
\code{mlr3pipelines::PipeOp}
}
\description{
Creates a PipeOp for impact encoding
}
