% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_urssaf_evaluation.R
\name{custom_eval_urssaf}
\alias{custom_eval_urssaf}
\title{Fonctions d'évaluation de la précision (spéciale URSSAF)}
\usage{
custom_eval_urssaf()
}
\value{
\code{MLsegmentr::eval_function}
}
\description{
Cette fonction construit un objet eval_function avec les fonctions
auxiliaires ci-dessus.
}
\details{
Etapes pour utiliser cette objet:
1- Entraînement du modèle. Déterminer les seuils F1 et F2 à partir de
l'échantillon d'entraînement.
2- faire un objet assesseur, avec cet objet EvaluationFunction,
3 -définir  comme segment la colonne interessant urssaf (6 segments)
4- Définir comme target l'anticipation d'un débit à 18 mois
}
