% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cv_task.R
\name{evaluate.cv_task}
\alias{evaluate.cv_task}
\title{Evaluate a cross-validated task}
\usage{
\method{evaluate}{cv_task}(task,
  eval_function = MLsegmentr::eval_precision_recall(),
  data_name = c("validation_data"), plot = TRUE,
  prediction_names = "score", target_names = "outcome",
  segment_names = NULL, remove_strong_signals = TRUE)
}
\description{
@inheritParams evaluate.sf_task
}
\details{
@return task with evaluation on first level. (no individual
cross-validated fold evaluation)
@export
}
