% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{evaluate}
\alias{evaluate}
\title{Évaluation du modèle}
\usage{
evaluate(
  ...,
  measures = get_default_measure(),
  should_remove_strong_signals = TRUE
)
}
\arguments{
\item{...}{\code{tasks} \cr Tasks to be evaluated.}

\item{measures}{\code{mlr3::Measure} Vecteur de mesures utilisées pour évaluer
la perforlance.}

\item{should_remove_strong_signals}{:: \code{logical(1)}\cr Faut-il retirer des
échantillons de test les entrerprises qui présentent des
signaux forts, c'est-à-dire 3 mois de défaut, ou une procédure collective
en cours ? Nécessite que les données contenues dans
\code{task[["hist_data"]]} possèdent le champs "time_til_outcome".}
}
\value{
\code{task} donnée en entrée à laquelle s'est ajoutée (ou a été
remplacé) un champs "model_performance", avec le résultat de la fonction
d'évaluation
}
\description{
Evalue les prédictions d'un ou plusieurs tasks, et retourne la performance
calculée sur l'échantillon (ou les échantillons dans le cadre de la
validatio croisée) de test.
}
