% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{explain}
\alias{explain}
\title{Explains model results}
\usage{
explain(task, ...)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}
}
\description{
Explains model results
}
