% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{explain_gam}
\alias{explain_gam}
\title{Get explanations for a gam model}
\usage{
explain_gam(task, data_to_explain)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{data_to_explain}{\code{data.frame}}
}
\value{

}
\description{
Get explanations for a gam model
}
