% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sf_task.R
\name{export.sf_task}
\alias{export.sf_task}
\title{Export de données}
\usage{
\method{export}{sf_task}(task, export_type, ...)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{export_type}{\code{"csv" | "mongodb"} \cr Export en csv, ou en
mongodb ?}

\item{...}{additional parameters for export functions.}
}
\value{
\code{sf_task} \cr L'objet \code{task} donné en entrée.
}
\description{
Exporte les données.
}
