% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{export_fiche_visite}
\alias{export_fiche_visite}
\title{Export d'une fiche visite}
\usage{
export_fiche_visite(sirets, database = "test_signauxfaibles",
  collection = "Features", batch, with_urssaf = FALSE,
  absolute_path = rprojroot::find_package_root_file("..", "output",
  "Fiches", batch))
}
\arguments{
\item{sirets}{Vecteur de sirets des établissements dont il faut exporter la
fiche}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{batch}{\code{character(1)} Choix du batch pour lequel se fait l'export}

\item{with_urssaf}{\code{logical(1)}\cr Si TRUE, les informations du montant des
dettes URSSAF sont incluses à la fiche.}

\item{absolute_path}{\code{character(1)} \cr Chemin d'accès absolu au dossier
d'export.}
}
\value{
\code{NULL}
Exporte la ou les fiches visites sous le nom
"Fiche_visite_{raison_sociale}.pdf"
}
\description{
Export d'une fiche visite
}
