% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{export_scores_to_csv}
\alias{export_scores_to_csv}
\title{Export des scores dans un tableau csv}
\usage{
export_scores_to_csv(formatted_data, algo, batch, absolute_path)
}
\arguments{
\item{formatted_data}{\code{data.frame()} \cr Données à exporter. Typiquement,
le data.frame retourné par la fonction \code{\link{format_for_export}}.}

\item{algo}{\code{character(1)} \cr Nom de l'algo qui figurera dans les objets
exportés}

\item{batch}{\code{character(1)} \cr Nom du batch qui figurera dans les objets
exportés}

\item{absolute_path}{\code{character(1)} \cr Chemin relatif du dossier dans lequel effectuer
l'export, par rapport à la racine du paquet R.}
}
\value{
\code{TRUE}. \cr Exporte un fichier csv, dans le dossier spécifié
en entrée, nommé automatiquement "aaaa_mm_jj_export_{algo}_{batch}" avec
éventuellement un suffixe "_vX" pour ne pas écraser de fichier existant.
}
\description{
Exporte les scores vers un tableau csv à partir des données formattées par la fonction
\code{\link{format_for_export}}.
}
