% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prediction_correction.R
\name{fetch_conj_data}
\alias{fetch_conj_data}
\title{Fonction de récupération des données de conjoncture Banque de France}
\usage{
fetch_conj_data()
}
\value{
\code{data.frame} Quelques séries de données de conjoncture de la Banque
de France, depuis le premier janvier 2014.
}
\description{
Via l'api webstat.
}
\details{
La variable webstat_client_ID est indispensable pour s'identifier auprès de
l'api webstat pour faire fonctionner w_series_list. Cf la documentation
officielle ou la page gollum "Accès extérieurs depuis labtenant dans R".
}
