% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_lgb.R
\name{full_light_gradient_boosting}
\alias{full_light_gradient_boosting}
\title{Quickly run the model -- temporary function}
\usage{
full_light_gradient_boosting(sample_size, batch)
}
\arguments{
\item{sample_size}{\code{integer()} \cr Taille de l'échantillon sur lequel
l'algorithme est entraîné.}

\item{batch}{\code{character()} \cr Nom du batch pour lequel le modèle est
entraîné et les dernières données sont sélectionnées.

@return \code{TRUE}
@export}
}
\description{
Charge les données, les prépare, entraîne l'algorithme, charge les
dernières données, prédit sur les dernières données et exporte les
résultats en un appel de fonction.
}
