% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{generate_failure_data}
\alias{generate_failure_data}
\title{Génère des données de défaillance factices}
\usage{
generate_failure_data(
  aggregate_by_n_months = TRUE,
  n_months = 3,
  add_random_predictions = TRUE,
  from = as.Date("2015-02-01"),
  to = as.Date("2015-12-01")
)
}
\arguments{
\item{aggregate_by_n_months}{\code{logical(1)} Faut-il des données agrégées par plages de
plusieurs mois ?}

\item{n_months}{Si\code{aggregate_by_n_months}, combien de mois dure la phase
d'agrégation (6 = semestre, 3 = trimestre, etc.)}

\item{add_random_predictions}{\code{logical(1)} Adds random predictions in a
"prediction" column}

\item{from}{\code{Date} Début de la période d'observation}

\item{to}{\code{Date} fin de la période d'observation}
}
\value{
\code{data.frame} avec deux secteurs A et B dans la colonne secteur, une
colonne "periode", des colonnes "count", "count_new_outcome" et
"prop_new_outcome", qui simulent respectivement: le nombre d'entreprises
dans le secteur, les nouvelles défaillances dans la plage temporelle, et la
proportion que cela représente.
}
\description{
Génère des données de défaillance factices
}
