% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sector_correction.R
\name{get_ape_to_naf}
\alias{get_ape_to_naf}
\title{Récupère la table de correspondance de ape à naf}
\usage{
get_ape_to_naf(path = file.path("raw_data", "naf5to1.csv"))
}
\arguments{
\item{path}{:: chemin d'accès}
}
\value{
\code{data.frame} table de correspondance avec colonnes "code_ape" et
"code_naf"
}
\description{
Cette table de correspondance doit par défaut se trouver dans le répertoire
"raw_data" et porter le nom "naf5to1.csv".
}
\details{
Ce fichier est disponible dans le répertoire "gollum", sur la page
"Ressources externes depuis labtenant dans R"
}
