% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sector_correction.R
\name{get_ape_to_secteur}
\alias{get_ape_to_secteur}
\title{Recupère la table de correspondance code_ape <-> secteur}
\usage{
get_ape_to_secteur(
  path = file.path("raw_data", "RWEBSTATS_table_passage.xlsx"),
  sheet = 6
)
}
\arguments{
\item{path}{:: chemin d'accès au fichier excel. Par défaut:
"raw_data/RWEBSTATS_table_passage.xlsx". Le fichier est disponible dans
"gollum", sur la page "Ressources externes depuis labtenant dans R".}

\item{sheet}{:: Feuille à récupérer, par défaut, la sixième}
}
\value{
\code{data.frame} table de correspondance avec les colonnes "code_ape", "secteur",
"nom_secteur".
}
\description{
Les secteurs sont utilisés dans les enquêtes de conjoncture de la Banque de
France
}
