% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{get_default_explanation_method}
\alias{get_default_explanation_method}
\title{Get default explanation method}
\usage{
get_default_explanation_method(task, data_to_explain)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{data_to_explain}{\code{data.frame}}
}
\value{
\code{function}
}
\description{
Get default explanation method
}
