% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_optimize.R
\name{get_default_param_set}
\alias{get_default_param_set}
\title{Default xgboost parameters to be tuned}
\usage{
get_default_param_set()
}
\value{
a paradox::ParamSet
}
\description{
This gives a default set of parameters to be tuned for the xgboost model.
To be used with \code{optimize_hyperparameters}.
}
