% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{get_fields}
\alias{get_fields}
\title{Get a list of field names}
\usage{
get_fields(training)
}
\arguments{
\item{training}{\code{logical()} \cr Si \code{TRUE}, uniquement des champs utilisés
pour l'entraînement sont retournés.}
}
\value{
\code{character()} \cr Vecteur de noms de variables
}
\description{
Fonction utilitaire pour récupérer rapidement une liste de noms de
variables (champs mongodb).  Utilitary function to quickly get a field name
list. Put training = TRUE to get only fields used for training. Put an
input to 0 to remove the related fields, to 1 to keep basic fields, and to
2 to include all fields.
}
