% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{get_fields_training_light}
\alias{get_fields_training_light}
\title{Obtenir une liste allégée de champs à exporter}
\usage{
get_fields_training_light()
}
\description{
TODO temporaire, retravailler get_fields.
task e58fb5d5-4bc1-410b-8287-c78e4fd442d0
}
