% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{get_gam_from_task}
\alias{get_gam_from_task}
\title{Fetches the trained gam model inside a task}
\usage{
get_gam_from_task(task)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}
}
\value{
\code{mgcv::gam} gam model
}
\description{
Fetches the trained gam model inside a task
}
