% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{get_gam_learner}
\alias{get_gam_learner}
\title{Get a generalized additive model learner}
\usage{
get_gam_learner()
}
\value{
\code{mlr3::Learner}
}
\description{
Get a generalized additive model learner
}
