% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{get_last_batch}
\alias{get_last_batch}
\title{Récupère les dernières données disponibles pour un batch}
\usage{
get_last_batch(last_batch, periods, database, collection, fields,
  min_effectif, rollback_months)
}
\arguments{
\item{last_batch}{\code{character(1)} \cr Batch auquel doit être importées les
données. Les modifications opérées par les batchs ultérieurs sont
ignorées.}

\item{periods}{\code{[Date()]} \cr Périodes d'intérêt, auquels charger les
données. Des périodes supplémentairs peuvent être chargées selon la
valeur de rollback_months.}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{fields}{\code{character()} \cr Noms des champs à requêter dans la base de
données. Doit contenir "siret" et "periode". Si égal à \code{NULL}, alors
charge tous les champs disponibles.}

\item{min_effectif}{\code{integer(1)} \cr Limite basse du filtrage de l'effectif
(la limite est incluse)}

\item{rollback_months}{\code{integer(1)}\cr Nombre de mois précédant le premier mois de
\code{periods} à charger. Permet d'effectuer des calculs de différences ou de
moyennes glissantes pour les périodes d'intérêt.}
}
\value{
\code{data.frame()} \cr
Données avec les colonnes décrites dans \code{fields}, pour les périodes
définies par \code{periods} et \code{rollback_months}
}
\description{
Cette fonction permet d'accéder rapidement aux dernières données
disponibles.
}
