% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{get_prepared_data}
\alias{get_prepared_data}
\title{Apply preparation pipeline and inspect prepared data}
\usage{
get_prepared_data(task, data)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{data}{\code{data.frame} données à préparer}
}
\value{
\code{data.frame} données passées en entrée après la préparation
}
\description{
Applique la pipeline de préparation sur les données d'entraînement ou de
test (que le premier échantillon en cas de validation croisée).
}
\details{
Cette fonction est uniquement prévue pour l'inspection du bon
fonctionnement de la pipeline de préparation.

L'objet task doit avoir une propriété "mlr3pipeline" de type
\code{mlr3pipelines::PipeOp} ou \code{mlr3pipelines::Graph}
}
