% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{get_scores}
\alias{get_scores}
\title{Récupération des scores}
\usage{
get_scores(database = "test_signauxfaibles", collection = "Scores",
  algo = "algo", method = "first", batchs = NULL, sirets = NULL)
}
\arguments{
\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{method}{\code{character(1)}\cr Ou bien "first" ou bien "last". See details.}

\item{algo`character(1)`\cr}{Algorithm from which to take the score. Defaults to "algo".}

\item{siret}{\code{character()} \cr Vecteur de sirets à requêter}
}
\value{
Returns a dataframe with columns "siret", "periode", "score"
(failure probability), "batch" (batch of the model used), "algo" (name of
the algorithm)
}
\description{
Récupération des scores
}
\section{Details}{

Si la méthode est "first", alors le dernier score disponible pour le
\strong{premier} batch disponible pour cette période est récupéré. Sinon, le
dernier score disponible pour le \strong{dernier} batch disponible est récupéré.
"first" représente plus fidèlement les scores qui ont été obtenues par le
passé, là où "last" représente les scores qui auraient été obtenues avec
le dernier algorithme en date (même si ce n'est pas toujours le cas, ça
dépend des périodes qui ont été  exportées avec cet algorithme).
}

