% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{get_test_task}
\alias{get_test_task}
\title{Get a fake task}
\usage{
get_test_task(
  fake_data = NULL,
  fake_target = "target",
  training_fields = "feature",
  stage = "prepare",
  resampling_strategy = "holdout",
  processing_pipeline = mlr3pipelines::PipeOpNOP$new(),
  learner = mlr3::LearnerClassifFeatureless$new(),
  measures = mlr3::msr("classif.acc")
)
}
\arguments{
\item{fake_data}{Data that is stored as "hist_data". If none is provided,
some data will be automatically set.}

\item{fake_target}{\code{character(1)} name of the column that should be use as
target for the learning.}

\item{training_fields}{\code{character()} name of the feature columns to learn
from}

\item{stage}{\code{character(1)} \cr Either "load", "split", "prepare", "train"
and "evaluate". The output task will be at the end of the specified stage.}

\item{resampling_strategy}{\code{character(1)} Either "holdout" or "cv". The
resampling strategy of the "train" stage.}

\item{processing_pipeline}{\code{mlr3pipelines::PipeOp} A preparation pipeline
used for the "prepare" stage.}

\item{learner}{\code{mlr3::Learner} A learner used at the "train" stage.}

\item{measures}{\code{mlr3::Measure} A list of measures used at the "evaluate"
stage.}
}
\value{
An sf_task, at the specified "stage"
}
\description{
Get a fake task
}
