% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{import_data}
\alias{import_data}
\title{Connexion à la base de donnée}
\usage{
import_data(
  database,
  collection,
  mongodb_uri,
  batch,
  min_effectif,
  date_inf = NULL,
  date_sup = NULL,
  fields = NULL,
  sirets = NULL,
  code_ape = NULL,
  subsample = NULL,
  verbose = FALSE,
  replace_missing = NULL,
  debug = FALSE
)
}
\arguments{
\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{mongodb_uri}{\code{character(1)} \cr url to the database in mongodb uri
format.}

\item{batch}{\code{character(1)} \cr Batch auquel doit être importées les
données. Les modifications opérées par les batchs ultérieurs sont
ignorées.}

\item{min_effectif}{\code{integer(1)} \cr Limite basse du filtrage de l'effectif
(la limite est incluse)}

\item{date_inf}{\code{Date(1)} \cr Limite inférieure de la période de temps
requêtée}

\item{date_sup}{\code{Date(1)} \cr Limite supérieure de la période de temps requêtée}

\item{fields}{\code{character()} \cr Noms des champs à requêter dans la base de
données. Doit contenir "siret" et "periode". Si égal à \code{NULL}, alors
charge tous les champs disponibles.}

\item{sirets}{\code{character()} \cr Liste de sirens à exporter. Si égale à
\code{NULL}, charge tous les sirens disponibles.}

\item{code_ape}{\code{character()} \cr Liste de code NAF ou APE (niveau 2 à 5) à
exporter. Si égale à \code{NULL}, charge tous les codes disponibles. Il est
permis de mélanger des codes de différents niveaux.}

\item{subsample}{\code{integer(1)} \cr Nombre d'objets (c'est-à-dire de couples
siret x periode) à échantillonner.}

\item{verbose}{\code{logical(1)} \cr Faut-il afficher dans le terminal des
informations sur l'évolution du chargement des données?}

\item{replace_missing}{\code{list()} \cr Liste nommée, dont les noms sont les
noms de variables et les valeurs sont les valeurs de remplacement des NA.
Si égal à NULL, alors des remplacements par défauts}
}
\value{
\code{data.frame()}
}
\description{
\code{import_data} permet de requêter des données mongoDB pour en
faire un dataframe ou un Spark dataframe. \cr
\code{factor_query} permet de fabriquer la requête d'aggrégation
correspondante. \cr
}
\section{Remplacement des valeurs manquantes par défaut}{

replace_missing <- list(
montant_part_patronale         = 0,
montant_part_ouvriere          = 0,
montant_echeancier             = 0,
ratio_dette                    = 0,
ratio_dette_moy12m             = 0,
montant_part_patronale_past_1  = 0,
montant_part_ouvriere_past_1   = 0,
montant_part_patronale_past_2  = 0,
montant_part_ouvriere_past_2   = 0,
montant_part_patronale_past_3  = 0,
montant_part_ouvriere_past_3   = 0,
montant_part_patronale_past_6  = 0,
montant_part_ouvriere_past_6   = 0,
montant_part_patronale_past_12 = 0,
montant_part_ouvriere_past_12  = 0,
apart_heures_consommees        = 0,
apart_heures_autorisees        = 0,
apart_entreprise               = 0,
tag_default                    = FALSE,
tag_failure                    = FALSE,
tag_outcome                    = FALSE
)
}

