% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cv_task.R
\name{join_for_evaluation.cv_task}
\alias{join_for_evaluation.cv_task}
\title{Join for evaluation}
\usage{
\method{join_for_evaluation}{cv_task}(task, ..., data_name)
}
\arguments{
\item{task}{\code{cv_task}}

\item{...}{\code{cv_task} \cr other \code{cv_task} to be joined}
}
\value{
cv_task
}
\description{
Join for evaluation
}
