% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sf_task.R
\name{join_for_evaluation.sf_task}
\alias{join_for_evaluation.sf_task}
\title{Joins tasks for a comparative evaluation}
\usage{
\method{join_for_evaluation}{sf_task}(task, ..., data_name)
}
\arguments{
\item{...}{\code{sf_task()} \cr One or several \code{sf_task}.}

\item{data_name}{\code{character(1)} \cr Name of data on which to evaluate.}
}
\value{
Data frame ready to use with evaluate
}
\description{
Joins tasks for a comparative evaluation
}
