% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_evaluation.R
\name{join_frames}
\alias{join_frames}
\title{Join several frames}
\usage{
join_frames(...)
}
\arguments{
\item{...}{:: \code{data.frame} \cr Frames to join}
}
\value{
\code{data.frame} \cr Joined frame
}
\description{
Joins the frames by all fields that are present in all the frames.
}
