% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_load_h2o_object.R
\name{load_h2o_object}
\alias{load_h2o_object}
\title{Charge un fichier enregistré par save_h2o_object}
\usage{
load_h2o_object(object_name, extension = NULL, absolute_path,
  last = TRUE, file_name = "")
}
\arguments{
\item{object_name}{\code{character(1)}\cr Nom de l'objet, spécifié de manière
identique qu'au moment de l'enregistrement. Cf \code{[save_h2o_object]}}

\item{extension}{\code{character(1)}\cr
L'extension du fichier à charger, ou bien "model" (pour un
\code{H2OBinomialModel}) ou bien "temap" (pour une \code{list(H2OFrame)}).}

\item{absolute_path}{Chemin d'accès relatif par rapport à la racine du projet.}

\item{last}{\code{logical()} \cr Si TRUE, charge le dernier fichier sauvegardé
correspondant à \code{object_name}, \code{extension} et \code{absolute_path}. Sinon, il
faut spécifier le nom complet: \code{file_name}.}

\item{file_name}{\code{character(1)}\cr Nom complet et exact du fichier à charger. N'est pris en compte uniquement si last est FALSE.}
}
\value{
\code{H2OBinomialModel() | }list(H2OFrame)` \cr
depending on the loaded model.
}
\description{
Cette fonction charge un fichier enregistré avec \code{save_h2o_object}, ou bien avec la racine \code{name}, l'extension du fichier, et le paramètre \code{last = TRUE}(qui charge le dernier fichier disponible), ou bien avec le nom exact avec le paramètre \code{file_name} et \code{last = FALSE}
}
