% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_load_h2o_object.R
\name{load_H2OFrame_list}
\alias{load_H2OFrame_list}
\title{Chargement d'une liste de H2OFrames}
\usage{
load_H2OFrame_list(absolute_path, filename)
}
\arguments{
\item{absolute_path}{\code{character()}\cr Chemin absolu (sans nom de fichier)}

\item{filename}{\code{character()}\cr Nom complet du fichier à enregistrer}
}
\value{
\code{H2OBinomialModel}
}
\description{
Charge une liste de H2OFrames. Utilisé pour le target encoding.
}
