% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_load_h2o_object.R
\name{load_H2OModel}
\alias{load_H2OModel}
\title{Chargement d'un fichier de classe "H2OBinomialModel"}
\usage{
load_H2OModel(absolute_path, filename)
}
\arguments{
\item{absolute_path}{\code{character(1)}\cr Chemin d'accès}

\item{filename}{Nom exact du fichier à charger.}
}
\value{
\code{list(H2OFrame)}
}
\description{
Charge un fichier de classe "H2OBinomialModel".
}
