% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R, R/interface_generics.R
\name{load_hist_data.sf_task}
\alias{load_hist_data.sf_task}
\alias{load_hist_data}
\title{Chargement de données historiques}
\usage{
\method{load_hist_data}{sf_task}(
  task,
  batch,
  database = task[["database"]],
  collection = task[["collection"]],
  mongodb_uri = task[["mongodb_uri"]],
  subsample = NULL,
  fields = get_fields(training = FALSE),
  date_inf = as.Date("2015-01-01"),
  date_sup = as.Date("2017-01-01"),
  min_effectif = 10L,
  siren = NULL,
  code_ape = NULL,
  debug = FALSE,
  ...
)

load_hist_data(task, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{batch}{\code{character(1)} \cr Batch auquel doit être importées les
données. Les modifications opérées par les batchs ultérieurs sont
ignorées.}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{mongodb_uri}{\code{character(1)} \cr url to the database in mongodb uri format.}

\item{subsample}{\code{integer(1)} \cr Nombre d'objets (c'est-à-dire de couples
siret x periode) à échantillonner.}

\item{fields}{\code{character()} \cr Noms des champs à requêter dans la base de
données. Doit contenir "siret" et "periode". Si égal à \code{NULL}, alors
charge tous les champs disponibles.}

\item{date_inf}{\code{Date} \cr Date inférieure du chargement des données.}

\item{date_sup}{\code{Date} \cr Date supérieure (exclue) du chargement des
données.}

\item{min_effectif}{\code{integer(1)} \cr Limite basse du filtrage de l'effectif
(la limite est incluse)}

\item{siren}{\code{character()} \cr Liste de sirens à exporter. Si égale à
\code{NULL}, charge tous les sirens disponibles.}

\item{code_ape}{\code{character()} \cr Liste de codes APE à exporter. Si égale à
\code{NULL}, charge tous les codes disponibles.}

\item{...}{Additional parameters depending on class of task}
}
\value{
\verb{[sf_task]} \cr
L'objet \code{task} donné en entrée auquel le champs "hist_data" a été
ajouté (ou écrasé), contenant un  data.frame() avec les colonnes incluses dans le paramètre d'entrée
\code{fields}, et pour chaque ligne un couple unique siret x periode.
}
\description{
Charge les données historiques de signaux faibles et les stocke dans un
champ "hist_data" de l'objet \code{task}
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

