% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R, R/interface_generics.R
\name{load_new_data.sf_task}
\alias{load_new_data.sf_task}
\alias{load_new_data}
\title{Chargement de nouvelles données}
\usage{
\method{load_new_data}{sf_task}(
  task,
  periods,
  batch,
  database = task[["database"]],
  collection = task[["collection"]],
  mongodb_uri = task[["mongodb_uri"]],
  fields = get_fields(training = FALSE),
  min_effectif = 10L,
  rollback_months = 1L,
  database_query_fun = query_database,
  debug = FALSE,
  ...
)

load_new_data(task, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{periods}{\verb{[Date()]} \cr Périodes d'intérêt, auquels charger les
données. Des périodes supplémentairs peuvent être chargées selon la
valeur de rollback_months.}

\item{batch}{\code{character(1)} \cr Batch auquel doit être importées les
données. Les modifications opérées par les batchs ultérieurs sont
ignorées.}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{mongodb_uri}{\code{character(1)} \cr url to the database in mongodb uri
format.}

\item{fields}{\code{character()} \cr Noms des champs à requêter dans la base de
données. Doit contenir "siret" et "periode". Si égal à \code{NULL}, alors
charge tous les champs disponibles.}

\item{min_effectif}{\code{integer(1)} \cr Limite basse du filtrage de l'effectif
(la limite est incluse)}

\item{rollback_months}{\code{integer(1)}\cr Nombre de mois précédant le premier
mois de \code{periods} à charger. Permet d'effectuer des calculs de différences
ou de moyennes glissantes pour les périodes d'intérêt.}

\item{database_query_fun}{\code{function} \cr Fonction qui permet de requêter la
base de données.}

\item{debug}{\code{logical(1)} \cr si \code{TRUE}, alors la requête est affichée.}

\item{...}{Additional parameters depending on class of task}
}
\value{
\verb{[sf_task]} \cr
L'objet \code{task} donné en entrée auquel le champs "new_data" a été
ajouté (ou écrasé), contenant un  data.frame() avec les colonnes incluses
dans le paramètre d'entrée \code{fields}, et pour chaque ligne un couple
unique siret x periode.
}
\description{
Chargement de nouvelles données

Loads new data
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

