% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sf_task.R
\name{log.sf_task}
\alias{log.sf_task}
\title{Log des expérimentations Machine Learning}
\usage{
\method{log}{sf_task}(task, database = task[["database"]],
  collection = "ml_logs", tracker = task[["tracker"]], ...)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)}. La collection vers laquelle exporter. Par
défaut, "ml_logs".}

\item{tracker}{\code{[mllogr::Tracker]} \cr Logger of ML experiment}
}
\value{
returns the \code{task} unchanged.
}
\description{
Permet de loguer dans une collection mongodb la nature et les différents
éléments de la tâche d'apprentissage. Cf le package \code{MLlogs}.
}
