% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{log_experiment}
\alias{log_experiment}
\title{Log a machine learning experiment}
\usage{
log_experiment(task, ...)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}
}
\description{
Log a machine learning experiment
}
