% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_log.R
\name{log_param}
\alias{log_param}
\title{Logs a parameter}
\usage{
log_param(task, key, value, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{key}{\code{character(1)} The key of the parameter to log.}

\item{value}{\code{any(1)} The value of the parameter to log.

Logs a parameter if the property "tracker" of \code{task} is not NULL.
Currently uses task[\link{"tracker"}] as a client for mlflow.}

\item{...}{Not used}
}
\description{
Logs a parameter
}
