% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{mark_known_sirets}
\alias{mark_known_sirets}
\title{Marks sirets from files as "known"}
\usage{
mark_known_sirets(df, full_paths)
}
\arguments{
\item{df}{\code{data.frame()} \cr Table avec au moins une colonne "siret"}

\item{full_paths}{\code{character()} \cr Chemin d'accès absolu}
}
\value{
\code{data.frame()} \cr Table donnée en entrée à laquelle a été apposé
une colonne "connu", qui vaut 1 si le siren correspondant a été trouvé dans
les fichiers spécifiés, 0 sinon.
}
\description{
Récupère des sirens d'une liste de fichier et ajoute une colonne à un
data.frame pour spécifier dans une colonne "connu" si le siren est présent dans l'un de ces
fichiers.
}
