% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\name{mongodb_connection}
\alias{mongodb_connection}
\title{Documentation des informations de connection à mongodb}
\arguments{
\item{mongodb_uri}{\code{character(1)} \cr url to the database in mongodb uri
format.}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}
}
\description{
Documentation des informations de connection à mongodb
}
