% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_file}
\alias{name_file}
\title{Nomme un fichier}
\usage{
name_file(relative_path, file_detail, file_extension = "",
  full_path = FALSE)
}
\arguments{
\item{relative_path}{\code{character(1)}\cr Chemin d'accès depuis
la racine
du projet.}

\item{file_detail}{\code{character(1)} \cr Description en un mot de la nature du
fichier sauvegardé}

\item{file_extension}{\code{character(1)} \cr Extension souhaitée pour le nom de
fichier}

\item{full_path}{\code{logical(1)} \cr Faut-il renvoyer le chemin complet
(\code{TRUE}) ou uniquement le nom de fichier (\code{FALSE})}
}
\value{
\code{character()} \cr
Un nom de fichier qui n'écrase pas les fichiers déjà contenues dans le
dossier spécifié par "relative_path", de la forme
"AAAA-MM-JJ_vX_{file_detail}.{file_extension}", où X est un chiffre
}
\description{
Nomme un fichier selon une convension précise afin de sauvegarder plusieurs
fichiers sans écraser les précédents.
}
