% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R, R/model_optimize.R
\name{optimize_hyperparameters}
\alias{optimize_hyperparameters}
\alias{optimize_hyperparameters.sf_task}
\title{Optimize hyperparameters}
\usage{
optimize_hyperparameters(task, ...)

\method{optimize_hyperparameters}{sf_task}(
  task,
  terminator = bbotk::trm("evals", n_evals = 10),
  param_set = get_default_param_set(),
  tuner = mlr3tuning::tnr("random_search"),
  measure = get_default_measure()[[1]],
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}

\item{terminator}{\code{bbotk::Terminator} Condition d'arrêt de l'optimisation
des hyperparamètres}

\item{param_set}{\code{paradox::ParamSet} Les paramètres à optimiser}

\item{tuner}{\code{mlr3tuning::Tuner} La méthode d'optimisation des
hyperparamètres}

\item{measure}{\code{mlr3::Measure} La métrique à optimiser}
}
\value{
La \code{task} donnée en entrée, auquel a été ajouté un champ
"model_parameters" avec les paramètres optimaux calculées.
}
\description{
Optimise les hyperparamètres du modèle. Nécessite des données
d'entraînement et de validation préparées.
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

