% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plots  a correlation graph}
\usage{
plot_correlation(task, fields)
}
\arguments{
\item{task}{Tâche d'apprentissage avec un champs "hist_data"}

\item{fields}{\code{character()} \cr
Sous-ensemble des variables pour lesquelles la corrélation doit être
affichée.}
}
\description{
Représente graphihquement la matrice de corrélation
}
