% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Affiche le graphique des importances relatives}
\usage{
plot_waterfall(frame)
}
\arguments{
\item{frame}{\code{data.table}}
}
\description{
Affiche le graphique des importances relatives
}
