% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{predict.sf_task}
\alias{predict.sf_task}
\title{Predict from a trained model}
\usage{
\method{predict}{sf_task}(object, data_names = c("new_data",
  "train_data", "validation_data", "test_data"),
  predict_fun = predict_model, ...)
}
\arguments{
\item{object}{\code{sf_task()} \cr Objet \code{sf_task} avec un modèle entraîné dans
le champs \code{model}. Les données doivent avoir été préparées.}

\item{data_names}{\code{character()} \cr Nom des données sur lesquelles prédire.}

\item{predict_fun}{\code{model, new_data -> predictions} \cr Prediction function}

\item{...}{Useless.}
}
\value{
\code{sf_task()}\cr
Task donnée en entrée, pour laquelle une colonne "score" a été rajoutée
aux données nommées dans \code{data_names}
}
\description{
Predict on some data.
Data should be prepared.
}
