% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{predict_model}
\alias{predict_model}
\title{Prédiction du model}
\usage{
predict_model(model, new_data)
}
\arguments{
\item{model}{\code{H2OModel()} \cr
Un modèle H2O.}

\item{new_data}{\code{H2OFrame()} \cr  Données au bon format pour alimenter le
modèle.}
}
\value{
\code{data.frame()} \cr
Un data.frame avec les colonnes "predicted_outcome" (\code{TRUE} ou \code{FALSE}),
"score" (entre 0 et 1), "siret" et "periode"
}
\description{
Produit une prédiction du modèle sur de nouvelles données. Renvoie le
résultat converti sous forme de \code{data.frame()}
}
