% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cv_task.R
\name{prepare.cv_task}
\alias{prepare.cv_task}
\title{Prepares a cross-validated task}
\usage{
\method{prepare}{cv_task}(task, data_names = c("train_data",
  "validation_data", "test_data", "new_data"))
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{data_names}{\code{character()} \cr Vecteur de noms des données à préparer.
Doivent-être des noms de champs valides de \code{task}.}
}
\value{
The task given as input where each fold-task has been prepared.
}
\description{
Prepares all subtasks (one for each fold) from cv_task.
}
