% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R, R/interface_sf_task.R
\name{prepare}
\alias{prepare}
\alias{prepare.sf_task}
\title{Prepare data}
\usage{
prepare(task, ...)

\method{prepare}{sf_task}(task, data_names = c("train_data",
  "validation_data", "test_data", "new_data"),
  tracker = task[["tracker"]])
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}

\item{data_names}{\code{character()} \cr Vecteur de noms des données à préparer.
Doivent-être des noms de champs valides de \code{task}.}

\item{tracker}{\code{[mllogr::Tracker]} \cr Logger of ML experiment}
}
\value{
\code{[sf_task]} \cr
L'objet \code{task} donné en entrée auquel les champs de données
préparées ont été ajoutées, avec une convention de nommage d'apposition
du préfixe "prepared_" aux noms des données noms préparées (par exemple:
"prepared_train_data" correspond aux données de "train_data" préparées).
}
\description{
Prépare les échantillons souhaités (listés dans \code{data_names}) pour
l'entraînement ou la prédiction de l'algorithme. Cf \code{[prepare_data]} pour
la nature de cette préparation.
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

