% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare_data_shape}
\alias{prepare_data_shape}
\title{Convertit la forme des données pour l'adapter aux exigences de l'algorithme}
\usage{
prepare_data_shape(data_to_prepare, training_fields)
}
\arguments{
\item{data_to_prepare}{:: \code{data.frame} \cr Données à préparer}

\item{training_fields}{\code{character()} \cr Les champs qui doivent être
conservées pour l'entraînement, après toutes les phases de préparation,
notamment le target encoding.}
}
\description{
1- Filtre les champs pour l'entraînement.
2- Convertit en matrice.
}
