% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare_frame}
\alias{prepare_frame}
\title{Préparation des données pour l'entraînement ou la prédiction}
\usage{
prepare_frame(data_to_prepare, test_or_train, te_map = NULL,
  save_or_load_map, outcome)
}
\arguments{
\item{data_to_prepare}{\code{data.frame()} \cr Données à préparer}

\item{test_or_train}{\code{"test" | "train"} \cr "train" if a new target
encoding map (\code{te_map}) should be computed; "test" otherwise}

\item{te_map}{\code{list(H2OFrame)}\cr Carte de calcul du target encoding. Utile
uniquement si \code{test_or_train} = "test", sinon elle est recalculée.}

\item{save_or_load_map}{\code{logical(1)} \cr Faut-il charger la carte de target
encoding (si \code{test_or_train}="test"), ou l'enregistrer (si \code{test_or_train}
= "train"). Chargement et sauvegarde se font avec le nom "te_map",
l'extension "te_map" et le chemin d'accès par défaut, cf
\code{[save_h2o_object]}.}

\item{outcome}{\code{character(1)}\cr Nom de la cible d'apprentissage sur
laquelle effectuer le target encoding.}
}
\value{
\code{list(2)} \cr
Une liste avec deux champs:
\itemize{
\item "data": \code{H2OFrame}; les données préparées
\item "te_map": \code{list(H2OFrame)}; la carte de target encoding
}
}
\description{
Inclut la conversion en H2O, et le target encoding.
La map pour le target encoding peut être sauvegardé, afin de permettre de
l'appliquer à de nouvelles données.
Le target encoding se fait avec du bruit pour éviter le surapprentissage.
}
