% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare_one_data_name}
\alias{prepare_one_data_name}
\title{Prepares one type of data (train, test, validation, new, other ...)}
\usage{
prepare_one_data_name(
  task,
  data_name,
  preparation_map_function,
  preparation_map_options,
  prepare_function,
  prepare_options,
  shape_frame_function,
  shape_frame_options
)
}
\description{
Prepares one type of data (train, test, validation, new, other ...)
}
