% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare_test_frame}
\alias{prepare_test_frame}
\title{Préparation des données pour la prédiction}
\usage{
prepare_test_frame(data_to_prepare, preparation_map, training_fields)
}
\arguments{
\item{data_to_prepare}{\code{data.frame()} \cr Données à préparer}

\item{preparation_map}{\code{list(H2OFrame)}\cr Carte de calcul du target encoding. Utile
uniquement si \code{test_or_train} = "test", sinon elle est recalculée.}

\item{training_fields}{\code{character()} \cr Les champs qui doivent être
conservées pour l'entraînement, après toutes les phases de préparation,
notamment le target encoding.}
}
\value{
\code{list(2)} \cr
Une liste avec deux champs:
\itemize{
\item "data": \code{H2OFrame}; les données préparées
\item "preparation_map": \code{list(H2OFrame)}; la carte de target encoding
}
}
\description{
Inclut la conversion en H2O, et le target encoding.
La map pour le target encoding peut être sauvegardé, afin de permettre de
l'appliquer à de nouvelles données.
Le target encoding se fait avec du bruit pour éviter le surapprentissage.
}
