% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare_test_frame_xgboost}
\alias{prepare_test_frame_xgboost}
\title{Préparation des données pour la prédiction d'un modèle xgboost}
\usage{
prepare_test_frame_xgboost(
  data_to_prepare,
  preparation_map,
  training_fields,
  outcome_field
)
}
\arguments{
\item{preparation_map}{\verb{?} \cr Carte de calcul du target encoding.}

\item{training_fields}{\code{character()} \cr Les champs qui doivent être
conservées pour l'entraînement, après toutes les phases de préparation,
notamment le target encoding.}
}
\value{
\code{list(2)} \cr
Une liste avec deux champs:
\itemize{
\item "data": \code{H2OFrame}; les données préparées
\item "preparation_map": \code{list(H2OFrame)}; la carte de target encoding
}
}
\description{
Inclut le target encoding.
La map pour le target encoding peut être sauvegardé, afin de permettre de
l'appliquer à de nouvelles données.
Le target encoding se fait avec du bruit pour éviter le surapprentissage.
}
