% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare_train_frame_linear}
\alias{prepare_train_frame_linear}
\title{Préparation des données pour l'entraînement d'un modèle linéaire}
\usage{
prepare_train_frame_linear(data_to_prepare, training_fields, outcome_field,
  target_encode_fields)
}
\arguments{
\item{data_to_prepare}{\code{data.frame()} \cr Données à préparer}

\item{training_fields}{\code{character()} \cr Les champs qui doivent être
conservées pour l'entraînement, après toutes les phases de préparation,
notamment le target encoding.}

\item{target_encode_fields}{\code{character()} \cr Nom des champs sur lesquels
effectuer du target encoding.}
}
\value{
\code{list(2)} \cr
Une liste avec deux champs:
\itemize{
\item "data": \code{H2OFrame}; les données préparées
\item "preparation_map": \code{list(H2OFrame)}; la carte de target encoding
}
}
\description{
Prépare les données pour l'entraînement, avec du target encoding et la
transformation en matrice pour alimenter xgboost.
}
\details{
La carte de target encoding est retourné en paramètre de sortie afin de
pouvoir l'utiliser sur d'autres échantillons de données.
}
