% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{print.sf_task}
\alias{print.sf_task}
\title{Print sf_task}
\usage{
\method{print}{sf_task}(x, ...)
}
\arguments{
\item{x}{\code{sf_task} \cr}

\item{...}{Useless}
}
\value{
invisible(x)
}
\description{
Print sf_task
}
