% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{read_aggregation_matrix}
\alias{read_aggregation_matrix}
\title{Lecture d'un fichier d'agrégation}
\usage{
read_aggregation_matrix(filepath)
}
\arguments{
\item{filepath}{\code{character()}\cr
Chemin d'accès du fichier d'agrégation à lire.}
}
\description{
Lecture d'un fichier d'agrégation
}
