% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_operations.R
\name{replace_hist_data}
\alias{replace_hist_data}
\alias{replace_new_data}
\title{Remplace les données historiques par d'autres données}
\usage{
replace_hist_data(task, data)

replace_new_data(task, data)
}
\arguments{
\item{task}{\code{sf_task}}

\item{data}{\code{data.frame()} Les données, doivent avoir des colonnes "siret",
"periode" et task[\link{"target"}]}
}
\value{
La sf_task d'origine

La sf_task d'origine
}
\description{
Reconstitue la tâche d'apprentissage task$mlr3task en conséquence.
}
\section{Functions}{
\itemize{
\item \code{replace_new_data}: 
}}

