% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_na}
\alias{replace_na}
\title{Remplacement de NA dans un data.frame}
\usage{
replace_na(frame, replacements_by_column, fail_if_column_missing = TRUE)
}
\arguments{
\item{frame}{\code{data.frame()} \cr
Le \code{data.frame()} dans lequel les NA doivent être remplacés.}

\item{replacements_by_column}{\code{list()} \cr Une liste nommée, dans laquelle
les noms sont des noms de colonnes, et les valeurs les valeurs de
remplacement pour ces colonnes.}

\item{fail_if_column_missing}{\code{logical(1)} \cr Si \code{true}, déclenche une
erreur si un nom de \code{replacements_by_column} ne figure pas dans la liste.}
}
\value{
La table \code{frame} dans laquelle les valeurs NA ont été remplacées
selon les indications de \code{replacements_by_column}
}
\description{
Remplace toutes les valeurs NA dans un data.frame par les valeurs spécifiés
colonne par colonne dans \code{replace_missing}
}
