% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sf_task.R
\name{save.sf_task}
\alias{save.sf_task}
\title{Saves model or map to disk}
\usage{
\method{save}{sf_task}(task, absolute_path, ...)
}
\arguments{
\item{absolute_path}{}
}
\description{
Saves model or map to disk
}
