% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_save_h2o_object.R
\name{save_h2o_object}
\alias{save_h2o_object}
\title{Enregistrer un objet H2O}
\usage{
save_h2o_object(object, object_name, absolute_path)
}
\arguments{
\item{object}{Objet à sauvegarder. Ou bien un "H2OBinomialModel", ou bien une liste de "H2OFrame"}

\item{object_name}{\code{character(1)} \cr Nom générique à donner au fichier, auxquels seront
automatiquement accolés une date et une version. Cf \code{[name_files]}.}

\item{absolute_path}{\code{character(1)} \cr Chemin relatif par rapport à la racine du projet R}
}
\value{
\code{character(1)}\cr
Chemin complet du fichier créé.
}
\description{
Enregistre un objet de type "H2OBinomialModel" ou "H2OFrame".
Attention, les modèles enregistrés ne sont pas compatibles d'une version
H2O sur l'autre.
}
