% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_save_h2o_object.R
\name{save_H2OFrame_list}
\alias{save_H2OFrame_list}
\title{Enregistre une liste de H2OFrames}
\usage{
save_H2OFrame_list(object, absolute_path, filename)
}
\arguments{
\item{object}{\code{list(H2OFrame)} \cr Liste de H2OFrames}

\item{absolute_path}{\code{character()}\cr Chemin absolu (sans nom de fichier)}

\item{filename}{\code{character()}\cr Nom complet du fichier à enregistrer}
}
\value{
\code{TRUE}
}
\description{
Enregistre une liste de \code{H2OFrame}.
}
