% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_save_h2o_object.R
\name{save_H2OModel}
\alias{save_H2OModel}
\title{Enregistrer un objet de classe "H2OBinomialModel"}
\usage{
save_H2OModel(object, absolute_path, filename)
}
\arguments{
\item{object}{\code{H2OBinomialModel}\cr Modèle H2O à enregistrer}

\item{absolute_path}{\code{character()}\cr Chemin absolu (sans nom de fichier)}

\item{filename}{\code{character()}\cr Nom complet du fichier à enregistrer}
}
\value{
\code{TRUE}
}
\description{
Enregistrer un objet de classe "H2OBinomialModel"
}
