% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{select_prepare_function}
\alias{select_prepare_function}
\title{Selectionne la fonction de préparation adéquate selon son intitulé}
\usage{
select_prepare_function(prepare_type, test_or_train)
}
\arguments{
\item{prepare_type}{\code{character(1)} \cr Intitulé de la fonction préparation.}

\item{test_or_train}{\code{"test" || "train"} \cr Fonction utilisée pour
préparer un échantillon d'entraînement ou de test ?}
}
\value{
Une fonction de préparation
}
\description{
Selectionne la fonction de préparation adéquate selon son intitulé
}
