% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_h2o_types}
\alias{set_h2o_types}
\title{Conversion des colonnes "string" en "enum" dans un H2OFrame}
\usage{
set_h2o_types(h2o_table)
}
\arguments{
\item{h2o_table}{Table h2o}
}
\value{
Table h2o avec les colonnes de type "string" remplacées par des colonnes de type "enum"
}
\description{
Conversion des colonnes "string" en "enum" dans un H2OFrame
}
