% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{set_verbose_level}
\alias{set_verbose_level}
\title{Active ou désactive le logging}
\usage{
set_verbose_level(task)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}
}
\value{
TRUE
}
\description{
Prend en compte l'attribut "verbose" de l'objet task pour fixer le bon
niveau de logging
}
