% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_theme.R
\name{sf_pal}
\alias{sf_pal}
\title{Return function to interpolate a sf color palette}
\usage{
sf_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a sf color palette
}
