% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_theme.R
\docType{data}
\name{sf_palettes}
\alias{sf_palettes}
\title{List of sf main palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
sf_palettes
}
\description{
List of sf main palettes
}
\keyword{datasets}
