% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{sf_task}
\alias{sf_task}
\title{Initialiser une tâche d'apprentissage}
\usage{
sf_task(
  mongodb_uri,
  database = "test_signauxfaibles",
  collection = "Features",
  id = "Signaux-faibles",
  target = "outcome",
  tracker = NULL
)
}
\arguments{
\item{mongodb_uri}{\code{character(1)} \cr url to the database in mongodb uri
format.}

\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{id}{\code{character(1)} Permet de donner un nom à la tâche d'apprentissage}

\item{tracker}{\code{mlflow::mlflow_run} \cr
Un mlflow_run pour tracker les modèles et expériences. Les logs
utiliseront la propriété \code{tracker$run_id}. Peut être laissé à "NULL" et
mlflow ne sera pas utilisé.}
}
\value{
\verb{[rsignauxfaibles::sf_task]} \cr
Un objet sf_task
}
\description{
Un objet s3 de type sf_task est défini, dans lequel seront défini et
stockés les tâches intermédiaires et les résultats de l'apprentissage.
}
