% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{shape_for_xgboost}
\alias{shape_for_xgboost}
\title{Returns a matrix for using xgboost}
\usage{
shape_for_xgboost(data_to_shape, options)
}
\arguments{
\item{data_to_shape}{\code{data.frame()}}

\item{options}{\code{list}(any)}
}
\value{
A matrix
}
\description{
Returns a matrix for using xgboost
}
