% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cv_task.R
\name{split_n_folds}
\alias{split_n_folds}
\title{Prepare cross-validation}
\usage{
split_n_folds(task, n_folds = 4, frac_test = 0.2,
  tracker = task[["tracker"]])
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{n_folds}{\code{integer(1)} \cr Number of folds}

\item{frac_test}{\code{numeric(1)} \cr Fraction of data which should be kept
aside}

\item{tracker}{\code{[mllogr::MLLogger]} \cr Logger of ML experiment}
}
\value{
\code{c(cv_task, sf_task)} \cr
A cross validated task with field "cross_validation", which is a list of
tasks, one for each fold.
}
\description{
Splits the data into \code{n_folds} for cross validation
}
