% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R
\name{split_snapshot_rdm_month}
\alias{split_snapshot_rdm_month}
\title{Découper l'échantillon en échantillon d'entraînement, de test et de validation}
\usage{
split_snapshot_rdm_month(data_sample, fracs, names, seed = 1234)
}
\arguments{
\item{data_sample}{\code{dataframe()}\cr
Données à scinder avec des champs "periode" et "siret".}

\item{fracs}{\code{numeric()} \cr Fraction des différents sous-échantillons.
Toutes les fractions doivent être dans ]0,1] et leur somme doit être égale
à 1.}

\item{names}{\code{character()} \cr Vecteur de noms des sous-échantillons, qui doit être de même
longueur que les fractions, ou de longueur 1. Si de longueur 1, alors un
préfixe "_X", avec X la numérotation des échantillons, est automatiquement ajouté.}

\item{seed}{\code{integer(1)} \cr Seed pour assurer la reproductibilité des
opérations aléatoires.}
}
\value{
\code{list(data.frames(3))} \cr
Liste de trois data.frames: \code{train}, \code{validation} and \code{test}
avec les couples (siret x periode) des trois échantillons calculés.
}
\description{
Scinde les données en échantillon d'entraînement, de validation et de
test, selon les proportions souhaitées. S'assure que deux établissements de la même entreprise ne soient pas
a la fois dans deux échantillons différents pour éviter la fuite
d'information d'un échantillon vers l'autre.
}
\details{
La fraction de l'échantillon de test est calculée par
1 - frac_train - frac_val. (frac_train + frac_val) doit donc être inférieur
a 1.
}
