% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prediction_correction.R
\name{string_denominator}
\alias{string_denominator}
\title{Fonction utilitaire}
\usage{
string_denominator(strings)
}
\arguments{
\item{strings}{\code{character()}  Vecteur de codes de séries de l'API webstat}
}
\description{
Prend un vecteur de codes de séries de l'API webstat de la Banque de France
et conserve les paramètres communs à toutes les séries. Les paramètres
différents sont remplacés par "*", dans un format utilisable par
\code{rwebstat::w_data}
}
