% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{subset_data_names_in_task}
\alias{subset_data_names_in_task}
\title{Keeps only the data_names that are indeed present in the task, warns if some
are missing.}
\usage{
subset_data_names_in_task(data_names, task)
}
\description{
Keeps only the data_names that are indeed present in the task, warns if some
are missing.
}
