% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_cv_task.R
\name{train.cv_task}
\alias{train.cv_task}
\title{Train a model on a cross-validated task}
\usage{
\method{train}{cv_task}(task, fields = get_fields(training = TRUE),
  parameters = NULL, seed = 123)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{fields}{\code{character()} \cr Liste des variables pour l'entraînement. Cf
\code{[get_fields]} pour les variables par défaut.}

\item{parameters}{\code{list(any())} \cr Paramètres du modèle. Cf
\code{[train_light_gradient_boosting]}. Si \code{NULL}, alors des paramètres par
défaut sont sélectionnés.}

\item{seed}{\code{integer()}\cr Graîne pour que les calculs aléatoires soient
reproductibles.}
}
\value{
The input task, where each cv-subtask has a trained model
}
\description{
Train models for each subtask, i.e. for each cross-validation fold.
}
