% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{train}
\alias{train}
\title{Trains model on data}
\usage{
train(task, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}
}
\description{
Trains model on data
}
