% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{train_linear}
\alias{train_linear}
\title{Train a linear model}
\usage{
train_linear(
  train_data,
  outcome,
  validation_data = NULL,
  parameters,
  seed = 123
)
}
\arguments{
\item{train_data}{\code{data.frame} \cr données d'entraînement, sous la forme d'un H2OFrame}

\item{outcome}{\code{character(1)} \cr Nom de la variable qui sert de cible
d'apprentissage}

\item{validation_data}{\code{H2OFrame} \cr Données d'évaluation, sous la forme d'un H2OFrame}

\item{parameters}{\code{list()} \cr Paramètres de la régression logistique.}

\item{seed}{\code{integer(1)} \cr Graine aléatoire pour que les opérations
aléatoires soient reproductibles.}
}
\description{
Trains a linear regularized model.
}
