% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{transformation_map_apply}
\alias{transformation_map_apply}
\title{Applies Box-cox transformation map to features}
\usage{
transformation_map_apply(transformation_map, data_to_transform)
}
\arguments{
\item{transformation_map}{As outputed by transformation_map_create}

\item{data_to_transform}{data to be transformed}
}
\description{
Applies Box-cox transformation map to features
}
