% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{transformation_map_create}
\alias{transformation_map_create}
\title{Normalise les données}
\usage{
transformation_map_create(data_to_transform)
}
\arguments{
\item{data_to_transform}{\code{data.frame()} \cr Données à transformer.}
}
\description{
Sélectionne des transformations pour que les données approchent de la
normalité.
}
