% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{xgboost_global_explainer}
\alias{xgboost_global_explainer}
\title{Explains global variable importance for an xgboost model}
\usage{
xgboost_global_explainer(task, aggregation_matrix, group_name, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{aggregation_matrix}{\code{dataframe()} \cr
Table de correspondance qui possède une colonne "variable" et une colonne
correspondant au paramètre \code{group_name}. Si égal à NULL (défaut), alors les variables ne
sont pas aggrégées. Utiliser \code{read_aggregation_matrix()} pour lire la
table de corrsepondance.}

\item{group_name}{\code{character()} \cr Nom de la colonne de
\code{aggregation_matrix} qui sert pour l'agrégation.}

\item{...}{Not used}
}
\description{
Importance des variables d'un modèle xgboost, avec la possibilité
d'aggréger selon un matrice d'agrégation
}
