% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R, R/interface_sf_task.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.sf_task}
\title{evaluates a model}
\usage{
evaluate(task, ...)

\method{evaluate}{sf_task}(task, eval_function = NULL,
  data_name = c("validation_data"), plot = TRUE,
  prediction_names = grep("^score", names(task[[data_name]]), value = T),
  target_names = "outcome", segment_names = NULL,
  remove_strong_signals = TRUE, tracker = task[["tracker"]])
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}

\item{eval_function}{\code{MLsegmentr::eval_function()} \cr Objet s3
d'évaluation.}

\item{data_name}{\code{character(1)} \cr Sur quelles données évaluer ?}

\item{plot}{\code{logical(1)} \cr Faut-il tracer la figure avec la fonction de
l'\code{eval_function} (si disponible)}

\item{remove_strong_signals}{\code{logical(1)}\cr
Faut-il retirer des échantillons de test ou de
validation les entrerprises qui présentent des signaux forts, c'est-à-dire 3 mois de défaut, ou une
procédure collective en cours ? Nécessite que les données contenues dans
\code{task[["hist_data"]]} possèdent le champs "time_til_outcome".}

\item{tracker}{\code{[mllogr::Tracker]} \cr Logger of ML experiment}
}
\value{
\code{task} donnée en entrée à laquelle s'est ajoutée (ou a été
remplacé) un champs "model_performance", avec le résultat de la fonction
d'évaluation
}
\description{
Evalue le modèle stocké dans l'objet task
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

