% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explain.R
\name{explain.sf_task}
\alias{explain.sf_task}
\title{Explains a single prediction}
\usage{
\method{explain}{sf_task}(
  task,
  type,
  aggregation_matrix = NULL,
  group_name = NULL,
  data_to_explain = NULL,
  plot_waterfall = FALSE,
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{type}{\code{character()} \cr Ou bien "global", "local".}

\item{aggregation_matrix}{\code{data.frame()} \cr  Une matrice d'agrégation. Si
égal à NULL, alors les résultats ne sont pas agrégés. Utiliser \code{read_aggregation_matrix()} pour lire la
table de corrsepondance.}

\item{group_name}{\code{character(1)} \cr Le nom de la colonne à utiliser dans
le data.frame \code{aggregation_matrix}. Obligatoire si aggregation_matrix est
différent de NULL.}

\item{data_to_explain}{\code{data.frame()} \cr Si type = "local" uniquement, les
données (lignes de "hist_data" ou "new_data") à expliquer.}

\item{...}{Not used}
}
\description{
Explains a single prediction
}
