% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sf_task.R
\name{generic_task}
\alias{generic_task}
\title{Documentation tâche générique}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{tracker}{\code{[mllogr::MLLogger]} \cr Logger of ML experiment}
}
\description{
Documentation tâche générique
}
