% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R
\name{generic_task}
\alias{generic_task}
\title{Documentation tâche générique}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{tracker}{\verb{[mlflow::mlflow_run]} \cr Logger de l'expérience
d'apprentissage automatique.}

\item{...}{Not used}
}
\description{
Documentation tâche générique
}
