% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{get_prepared_data}
\alias{get_prepared_data}
\title{Apply preparation pipeline and inspect prepared data}
\usage{
get_prepared_data(task, train_or_test)
}
\description{
Apply preparation pipeline and inspect prepared data
}
