% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_export.R
\name{get_scores}
\alias{get_scores}
\title{Récupération des scores}
\usage{
get_scores(database = "test_signauxfaibles", collection = "Scores",
  mongodb_uri, algo = "algo", method = "last", batchs = NULL,
  sirets = NULL)
}
\arguments{
\item{database}{\code{character(1)} \cr Nom de la base de données vers laquelle
param exporter. Par défaut, celle stockée dans \code{task}.}

\item{collection}{\code{character(1)} \cr Nom de la collection vers laquelle
exporter. Par défaut, celle stockée dans \code{task}.}

\item{mongodb_uri}{\code{character(1)} \cr url to the database in mongodb uri format.}

\item{algo}{\code{character(1)}\cr Algorithm from which to take the score. Defaults to "algo".}

\item{method}{\code{character(1)}\cr Ou bien "first" ou bien "last". Se référer
à la section Details.}

\item{siret}{\code{character()} \cr Vecteur de sirets à requêter}
}
\value{
Retourne un \code{data.frame} avec des colonnes "siret", "periode", "score"
(probabilité de défaillance selon la définition signaux-faibles), "batch"
(Nom du batch d'intégration pour lequel le score a été calculé), "algo"
(nom de l'algorithme utilisé) et "timestamp" (date et heure d'intégration des scores
à la base de données)
}
\description{
Récupération des scores
}
\section{Details}{

Si la méthode est "first", alors le dernier score disponible pour le
\strong{premier} batch disponible pour cette période est récupéré. Sinon, le
dernier score disponible pour le \strong{dernier} batch disponible est récupéré.
"first" représente plus fidèlement les scores qui ont été obtenues par le
passé, là où "last" représente les scores qui auraient été obtenues avec
le dernier algorithme en date (même si ce n'est pas toujours le cas, ça
dépend des périodes qui ont été exportées avec cet algorithme).
}

