% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{get_test_task}
\alias{get_test_task}
\title{Get a fake task}
\usage{
get_test_task(
  fake_data = NULL,
  fake_target = "target",
  training_fields = "feature",
  stage = "prepare",
  resampling_strategy = "holdout",
  processing_pipeline = mlr3pipelines::PipeOpNOP$new(),
  learner = mlr3::LearnerClassifFeatureless$new(),
  measures = msr("classif.acc")
)
}
\arguments{
\item{fake_data}{Data that is stored as "hist_data". If none is provided,
some data will be automatically set.}

\item{stage}{\code{character(1)} \cr Either "load", "split" or "prepare". The
output task will be at the end of the specified stage.}
}
\description{
Get a fake task
}
