% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R, R/model_optimize.R
\name{optimize_hyperparameters}
\alias{optimize_hyperparameters}
\alias{optimize_hyperparameters.sf_task}
\title{Optimize hyperparameters}
\usage{
optimize_hyperparameters(task, ...)

\method{optimize_hyperparameters}{sf_task}(
  task,
  terminator = bbotk::trm("evals", n_evals = 10),
  param_set = get_default_param_set(),
  tuner = mlr3tuning::tnr("random_search"),
  measure = get_default_measure()[[1]],
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}

\item{fields}{\code{character()} \cr Liste des variables pour l'entraînement. Cf
\verb{[get_fields]} pour les variables par défaut.}

\item{n_init}{\code{integer(1)} \cr Nombre d'évaluations aléatoires initiales.}

\item{n_iter}{\code{integer(1)} \cr Nombre d'itérations (d'évaluations)
d'optimisation.}

\item{train_pipe}{\code{function} \cr Fonction d'entraînement et d'évaluation
compatible avec l'optimisation avec "mlrMBO" (donc avec des attributs
spécifiques), notamment comme celles créées avec
\verb{[smoof::makeSingleObjectiveFunction]}}

\item{optim_bounds}{\code{ParamSet} \cr Objet précisant l'espace exploré par
l'optimisation, tel que donné par \verb{[ParamHelpers::ParamSet]}. Si NULL,
alors des limites par défaut sont fixées.}
}
\value{
La \code{task} donnée en entrée, auquel a été ajouté un champ
"model_parameters" avec les paramètres optimaux calculées.
}
\description{
Optimise les hyperparamètres du modèle. Nécessite des données
d'entraînement et de validation préparées.
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

