% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_generics.R, R/interface_sf_task.R
\name{optimize_hyperparameters}
\alias{optimize_hyperparameters}
\alias{optimize_hyperparameters.sf_task}
\title{Optimize hyperparameters}
\usage{
optimize_hyperparameters(task, ...)

\method{optimize_hyperparameters}{sf_task}(task,
  fields = get_fields(training = TRUE), n_init = 6, n_iter = 12,
  train_pipe = NULL, optim_bounds = NULL)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{...}{Additional parameters depending on class of task}

\item{fields}{\code{character()} \cr Liste des variables pour l'entraînement. Cf
\code{[get_fields]} pour les variables par défaut.}

\item{n_init}{\code{integer(1)} \cr Nombre d'évaluations aléatoires initiales.}

\item{n_iter}{\code{integer(1)} \cr Nombre d'itérations (d'évaluations) d'optimisation.}

\item{train_pipe}{\code{function} \cr Fonction d'entraînement et d'évaluation
compatible avec l'optimisation avec "mlrMBO" (donc avec des attributs
spécifiques), notamment comme celles créées avec
\code{[smoof::makeSingleObjectiveFunction]}}

\item{optim_bounds}{\code{ParamSet} \cr Objet précisant l'espace exploré par
l'optimisation, tel que donné par \code{[ParamHelpers::ParamSet]}. Si NULL,
alors des limites par défaut sont fixées.}
}
\value{
La \code{task} donnée en entrée, auquel a été ajouté un champ
"model_parameters" avec les paramètres optimaux calculées.
}
\description{
Optimise les hyperparamètres du modèle. Nécessite des données
d'entraînement et de validation préparées.
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

