% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_predict.R
\name{predict.cv_task}
\alias{predict.cv_task}
\title{Predict model on new data for a cross-validated task}
\usage{
\method{predict}{cv_task}(object, data_names = c("new_data",
  "train_data", "validation_data", "test_data"),
  predict_fun = predict_model, ...)
}
\arguments{
\item{object}{\code{sf_task()} \cr Objet \code{sf_task} avec un modèle entraîné dans
le champs \code{model}. Les données doivent avoir été préparées.}

\item{data_names}{\code{character()} \cr Nom des données sur lesquelles prédire.}

\item{predict_fun}{\code{model, new_data -> predictions} \cr Prediction function}

\item{...}{Useless.}
}
\value{
A task where each cross-validated fold has predictions
}
\description{
Predict on new data for each cross-validated fold.
}
