% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare.cv_task}
\alias{prepare.cv_task}
\title{Prepares a cross-validated task}
\usage{
\method{prepare}{cv_task}(
  task,
  data_names = c("train_data", "test_data", "new_data"),
  preparation_map_function = create_fte_map,
  preparation_map_options = list(outcome_field = outcome_field, target_encode_fields =
    c("code_ape_niveau2", "code_ape_niveau3")),
  prepare_function = apply_fte_map,
  prepare_options = list(outcome_field = outcome_field, target_encode_fields =
    c("code_ape_niveau2", "code_ape_niveau3")),
  shape_frame_function = shape_for_xgboost,
  shape_frame_options = list(),
  outcome_field = "outcome",
  preprocessing_strategy = NULL,
  training_fields = get_fields(training = TRUE),
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{data_names}{\code{character()} \cr Vecteur de noms des données à préparer.
Doivent-être des noms de champs valides de \code{task}.}

\item{preprocessing_strategy}{character() \cr
Description en langage naturel de la stratégie de préparation des données
pour le logging de l'expérience.}

\item{training_fields}{\code{character()} \cr Les champs qui doivent être
conservées pour l'entraînement, après toutes les phases de préparation,
notamment le target encoding.}

\item{...}{Unused.}
}
\value{
The task given as input where each fold-task has been prepared.
}
\description{
Prepares all subtasks (one for each fold) from cv_task. Benefits from the
same defaults as for sf_tasks.
}
