% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare.cv_task}
\alias{prepare.cv_task}
\title{Prepares a cross-validated task}
\usage{
\method{prepare}{cv_task}(task, data_names = c("train_data",
  "validation_data", "test_data", "new_data"),
  prepare_train_function = "xgboost",
  prepare_test_function = "xgboost", outcome_field = "outcome",
  tracker = NULL, preprocessing_strategy = NULL,
  training_fields = get_fields(training = TRUE), ...)
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{data_names}{\code{character()} \cr Vecteur de noms des données à préparer.
Doivent-être des noms de champs valides de \code{task}.}

\item{prepare_train_function}{:: (data_to_prepare :: data.frame) ->
list(data :: data.frame, preparation_map :: any) \cr
A function that prepares the training data and gives a preparation_map to
be applied for further data preparation.
Can be equal to "xgboost" or "linear" for quick reference to the right
preparation function.}

\item{prepare_test_function}{:: (data_to_prepare :: data.frame,
preparation_map :: any) -> data.frame \cr
A function that uses the preparation map to prepare further data (test,
validation data etc.)
Can be equal to "xgboost" or "linear" for quick reference to the right
preparation function.}

\item{tracker}{\code{[mllogr::Tracker]} \cr Logger of ML experiment}

\item{preprocessing_strategy}{character() \cr
Description en langage naturel de la stratégie de préparation des données
pour le logging de l'expérience.}

\item{training_fields}{\code{character()} \cr Les champs qui doivent être
conservées pour l'entraînement, après toutes les phases de préparation,
notamment le target encoding.}

\item{...}{Unused.}
}
\value{
The task given as input where each fold-task has been prepared.
}
\description{
Prepares all subtasks (one for each fold) from cv_task. Benefits from the
same defaults as for sf_tasks.
}
