% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{prepare.sf_task}
\alias{prepare.sf_task}
\title{Préparation des échantillons pour l'entraînement et l'évaluation}
\usage{
\method{prepare}{sf_task}(
  task,
  training_fields = get_fields(training = TRUE),
  outcome_field = NULL,
  processing_pipeline = get_default_pipeline(),
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{training_fields}{\code{character()} \cr
Les champs qui doivent être conservées pour l'entraînement.}

\item{outcome_field}{\code{character(1)} \cr
Champ de la variable à prédire.}

\item{processing_pipeline}{\code{mlr3pipelines::PipeOp} \cr
Pipeline de préparation des données}

\item{...}{Unused.}
}
\value{
\verb{[sf_task]} \cr
L'objet \code{task} donné en entrée auquel les champs de données
préparées ont été ajoutées, avec une convention de nommage d'apposition
du préfixe "prepared_" aux noms des données préparées (par exemple:
"prepared_train_data" correspond aux données de "train_data" préparées).
}
\description{
Prépare les données pour l'entraînement ou la prédiction de l'algorithme.
}
