% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R, R/interface_generics.R
\name{split_data.sf_task}
\alias{split_data.sf_task}
\alias{split_data}
\title{Scission des données en échantillon d'entraînement, de validation et de
test.}
\usage{
\method{split_data}{sf_task}(
  task,
  fracs = c(0.6, 0.2, 0.2),
  names = c("train", "validation", "test"),
  tracker = task[["tracker"]],
  ...
)

split_data(task, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{fracs}{\code{numeric()} \cr Fraction des différents sous-échantillons.
Toutes les fractions doivent être dans ]0,1] et leur somme doit être égale
à 1.}

\item{names}{\code{character()} \cr Vecteur de noms des sous-échantillons, qui doit être de même
longueur que les fractions, ou de longueur 1. Si de longueur 1, alors un
préfixe "_X", avec X la numérotation des échantillons, est automatiquement ajouté.}

\item{tracker}{\verb{[mllogr::Tracker]} \cr Logger of ML experiment}

\item{...}{Not used}
}
\value{
\verb{[sf_task]} \cr
L'objet \code{task} donné en entrée auquel les champs "train_data",
"validation_data" et "test_data" ont été
ajoutés (ou écrasés), chacun contenant un data.frame() avec les colonnes
de \code{task[["hist_data"]]} et un sous-ensemble (possiblement vide) de ces
lignes.
}
\description{
Scinde les données historiques en échantillon d'entraînement, de validation
et de test, selon les proportions souhaitées. S'assure que deux
établissements de la même entreprise ne soient pas à la fois dans deux
échantillons différents pour éviter la fuite d'information d'un échantillon
vers l'autre.
}
\details{
La fraction de l'échantillon de test est calculée par
1 - frac_train - frac_val. (frac_train + frac_val) doit donc être inférieur
à 1. Le seul cas où cette condition n'est pas testée est lorsque frac_train
= 1.
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

