% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_split.R, R/interface_generics.R
\name{split_data.sf_task}
\alias{split_data.sf_task}
\alias{split_data}
\title{Scission des données en échantillon d'entraînement et de test.}
\usage{
\method{split_data}{sf_task}(task, ratio = 2/3, nfolds = 5, resampling_strategy = "holdout", ...)

split_data(task, ...)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{ratio}{\code{numeric(1)} \cr Ratio des données utilisées pour
l'entraînement. Ignoré si 'resampling_strategy != "holdout"'}

\item{nfolds}{\code{numeric(1)} \cr Nombre d'échantillons de validation croisé.
Ignoré si 'resampling_strategy != "cv"'.}

\item{resampling_strategy}{\code{character(1)} \cr Ou bien "holdout"
ou "cv". Toute autre valeur n'occasionne aucun échantillonnage.}

\item{...}{Not used}
}
\value{
\verb{[sf_task]} \cr L'objet \code{task} donné en entrée auquel les
champs "train_data", et "test_data" ont été ajoutés (ou écrasés), chacun
contenant un data.frame() avec les colonnes de \code{task[["hist_data"]]} et un
sous-ensemble (possiblement vide) de ses lignes.
}
\description{
Scinde les données historiques en échantillon d'entraînement et de test,
selon le ratio souhaité. S'assure que deux établissements de la
même entreprise ne soient pas à la fois dans deux échantillons différents
pour éviter la fuite d'information d'un échantillon vers l'autre.
}
\section{Methods (by class)}{
\itemize{
\item \code{sf_task}: 
}}

