% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{train.sf_task}
\alias{train.sf_task}
\title{Entraînement de l'algorithme}
\usage{
\method{train}{sf_task}(
  task,
  seed = 0,
  learner = get_default_learner(),
  store_models = TRUE,
  keep_intermediate_results = TRUE,
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{seed}{\code{integer()}\cr Graîne pour que les calculs aléatoires soient
reproductibles.}

\item{learner}{\code{mlr3::ClassifLearner} \cr
Un modèle de classification mlr3 (les paramètres sont directement
spécifiés dans le modèle).}

\item{store_models}{\code{logical(1)} \cr
Faut-il sauvegardé tous les modèles entraînés pour référence ultérieure ?}

\item{...}{Not used}

\item{keep_results}{\code{logical(1)} \cr
Faut-il conserver les états intermédiaires (données préparées)}
}
\value{
\verb{[sf_task]} \cr L'objet \code{task} donné en entré, auquel a été ajouté
(ou écrasé) le champs "mlr3resample_result" si les données ont été
échantillonnées, "mlr3model" sinon, dans lequel est stocké un modèle
compatible avec la fonction \verb{[predict.sf_task]}.
}
\description{
Entraîne un algorithme sur les données \code{task[["train_data"]]}. Cf
\verb{[train_light_gradient_boosting]} pour plus de détails sur le modèle
entraîné.
}
