% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{train.sf_task}
\alias{train.sf_task}
\title{Entraînement de l'algorithme}
\usage{
\method{train}{sf_task}(
  task,
  outcome = "outcome",
  parameters = NULL,
  seed = 0,
  train_fun = "xgboost",
  ...
)
}
\arguments{
\item{task}{\verb{[sf_task]} \cr Objet s3 de type sf_task}

\item{outcome}{\code{character()} \cr Nom de la colonne dans laquelle on trouve
l'objectif d'apprentissage.}

\item{parameters}{\code{list(any())} \cr Paramètres du modèle. Cf
\verb{[train_light_gradient_boosting]}. Si \code{NULL}, alors des paramètres par
défaut sont sélectionnés.}

\item{seed}{\code{integer()}\cr Graîne pour que les calculs aléatoires soient
reproductibles.}

\item{train_fun}{\code{training_data} \cr Training_function, or "xgboost" or
"linear"}

\item{...}{Not used}
}
\value{
\verb{[sf_task]} \cr L'objet \code{task} donné en entré, auquel a été ajouté
(ou écrasé) le champs "model", dans lequel est stocké un modèle compatible
avec la fonction \verb{[prepare.sf_task]}.
}
\description{
Entraîne un algorithme sur les données \code{task[["train_data"]]}. Cf
\verb{[train_light_gradient_boosting]} pour plus de détails sur le modèle
entraîné.
}
