% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_sf_task.R
\name{train.sf_task}
\alias{train.sf_task}
\title{Entraînement de l'algorithme}
\usage{
\method{train}{sf_task}(task, fields = get_fields(training = TRUE),
  parameters = NULL, seed = 123, tracker = task[["tracker"]])
}
\arguments{
\item{task}{\code{[sf_task]} \cr Objet s3 de type sf_task}

\item{fields}{\code{character()} \cr Liste des variables pour l'entraînement. Cf
\code{[get_fields]} pour les variables par défaut.}

\item{parameters}{\code{list(any())} \cr Paramètres du modèle. Cf
\code{[train_light_gradient_boosting]}. Si \code{NULL}, alors des paramètres par
défaut sont sélectionnés.}

\item{seed}{\code{integer()}\cr Graîne pour que les calculs aléatoires soient
reproductibles.}

\item{tracker}{\code{[mllogr::Tracker]} \cr Logger of ML experiment}
}
\value{
\code{[sf_task]} \cr L'objet \code{task} donné en entré, auquel a été ajouté
(ou écrasé) le champs "model", dans lequel est stocké un modèle compatible
avec la fonction \code{[prepare.sf_task]}.
}
\description{
Entraîne un algorithme sur les données \code{task[["train_data"]]}. Cf
\code{[train_light_gradient_boosting]} pour plus de détails sur le modèle
entraîné.
}
